Added detete and search methods are added

+++ src/main/java/com/tcs/dao/PatientDao.java
@@ -8,8 +8,8 @@ public abstract class PatientDao {
 	public abstract String validateLogin(String username, String password);
 	public abstract boolean patientRegistration(Patient patient);
 	public abstract boolean updatePatient(Patient patient);
-	public abstract boolean deletePatient(int id);
-	public abstract Patient searchPatient(int id);
+	public abstract boolean deletePatient(long id);
+	public abstract Patient searchPatient(long id);
 	public abstract Patient[] viewAllPatients();
 	public abstract Medicine[] viewMedcines(int id);
 	public abstract Diagnostic[] viewDiagnostics(int id);
